Rename FSconfig.h to options_file_system.h in MPLAB-X projects.
