+++ Microchip/Include/MDD-File-System/FSIO.h
@@ -39,7 +39,7 @@
 #define  FS_DOT_H
 
 #include "GenericTypeDefs.h"
-#include "FSconfig.h"
+#include "options_file_system.h"
 #include "MDD-File-System/FSDefs.h" 
 #include "stddef.h"
