+++ RollPitchYaw/RollPitchYaw-auav3.X/nbproject/configurations.xml
@@ -100,12 +100,12 @@
       <logicalFolder name="RollPitchYaw/Config"
                      displayName="RollPitchYaw/Config"
                      projectFiles="true">
-        <itemPath>../../RollPitchYaw/Config/FSconfig.h</itemPath>
         <itemPath>../../RollPitchYaw/Config/options.h</itemPath>
         <itemPath>../../Config/options_magnetometer.h</itemPath>
         <itemPath>../Config/options_osd.h</itemPath>
         <itemPath>../Config/options_ports.h</itemPath>
         <itemPath>../Config/options_nv_memory.h</itemPath>
+        <itemPath>../Config/options_file_system.h</itemPath>
       </logicalFolder>
     </logicalFolder>
     <logicalFolder name="LinkerScript"
