+++ libFlashFS/MDD_AT45D.h
@@ -24,7 +24,7 @@
 
 
 #include "GenericTypeDefs.h"
-#include "FSconfig.h"
+#include "options_file_system.h"
 #include "MDD-File-System/FSDefs.h"
 
 #define FALSE   0
